refactor: 重构 Header 组件
