+++ src/BootstrapBlazor.Shared/Components/Header.razor.js
@@ -1,0 +1,20 @@
+﻿import EventHandler from "../../../_content/BootstrapBlazor/modules/event-handler.js"
+
+export function init() {
+    const scrollTop = () => (document.documentElement && document.documentElement.scrollTop) || document.body.scrollTop
+    let prevScrollTop = 0;
+    EventHandler.on(document, 'scroll', () => {
+        const items = document.querySelectorAll('.navbar-header, .coms-search')
+        const currentScrollTop = scrollTop()
+        if (currentScrollTop > prevScrollTop) {
+            items.forEach(item => item.classList.add('hide'))
+        } else {
+            items.forEach(item => item.classList.remove('hide'))
+        }
+        prevScrollTop = currentScrollTop
+    })
+}
+
+export function dispose() {
+    EventHandler.off(document, 'scroll')
+}
